condaenv.mbo45om1.requirements.txt
accelerate==1.0.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
async-timeout==4.0.3
attrs==24.2.0
beartype==0.19.0
bleach==6.1.0
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
datasets==3.0.1
dill==0.3.8
docker-pycreds==0.4.0
einops==0.8.0
ema-pytorch==0.7.0
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.6.1
gitdb==4.0.11
gitpython==3.1.43
huggingface-hub==0.25.2
idna==3.10
imagen-pytorch==2.1.0
kaggle==1.6.17
kiwisolver==1.4.7
kornia==0.7.3
kornia-rs==0.1.5
matplotlib==3.6.2
multidict==6.1.0
multiprocess==0.70.16
packaging==24.1
pandas==2.2.3
pathtools==0.1.2
pillow==9.3.0
promise==2.3
propcache==0.2.0
protobuf==4.25.5
psutil==6.0.0
pyarrow==17.0.0
pydantic==2.9.2
pydantic-core==2.23.4
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytorch-warmup==0.1.1
pytz==2024.2
pyyaml==6.0.2
regex==2024.9.11
safetensors==0.4.5
sentencepiece==0.2.0
sentry-sdk==2.14.0
setproctitle==1.3.3
shortuuid==1.0.13
six==1.16.0
smmap==5.0.1
text-unidecode==1.3
tokenizers==0.20.1
tqdm==4.66.5
transformers==4.45.2
tzdata==2024.2
wandb==0.13.5
webencodings==0.5.1
xxhash==3.5.0
yarl==1.15.1